chatbot/requirements.txt
attrs==21.2.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
decorator==4.4.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipython==7.20.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
jmespath==0.10.0
lxml==4.6.3
numpy==1.20.3
opencv-contrib-python==4.5.2.54
parsel==1.6.0
parso==0.8.1
peewee==3.14.4
pickleshare==0.7.5
Pillow==8.2.0
pony==0.7.14
priority==1.3.0
prompt-toolkit==3.0.16
Protego==0.1.16
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pygame==2.0.0
Pygments==2.8.0
pyOpenSSL==20.0.1
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
Scrapy==2.5.0
service-identity==21.1.0
simple-draw==2.7.0
six==1.15.0
soupsieve==2.2.1
tabulate==0.8.9
termcolor==1.1.0
traitlets==5.0.5
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.4
vk-api==11.9.3
w3lib==1.22.0
wcwidth==0.2.5
zope.interface==5.4.0
